proj2/TeachersPetBot/requirements.txt
python-dotenv==0.19.0
discord.py==1.7.3
better_profanity==0.7.0
numpy==1.21.2
validators==0.18.2
discord-components==2.1.1
coverage==5.5
pytest==6.2.5
pylint==2.11.1
psutil==5.8.0
apscheduler==3.8.1
pyshorteners~=1.0.1
quickchart.io==1.0.0